+++ R/att_from_rmds.R
@@ -81,7 +81,7 @@ att_from_rmd <- function(path, temp_dir = tempdir(), warn = -1,
 
 #' @export
 att_from_rmds <- function(path = "vignettes",
-                          pattern = "*.[.](Rmd|rmd)$",
+                          pattern = "*.[.](Rmd|rmd|qmd)$",
                           recursive = TRUE, warn = -1,
                           inside_rmd = FALSE, inline = TRUE) {